got vehicle caching to work with reset functionality
